Cleanup removeMode
