+++ platform/cli/src/commands/removeMode.js
@@ -3,9 +3,8 @@ import chalk from 'chalk';
 
 import {
   uninstallNPMPackage,
-  readPluginConfigFile,
-  removeModeFromConfig,
-  writePluginConfigFile,
+  findOhifExtensionsToRemoveAfterRemovingMode,
+  removeModeFromConfigFile,
   validateModeYarnInfo,
   getYarnInfo,
 } from './utils/index.js';
@@ -13,77 +12,6 @@ import removeExtensions from './removeExtensions.js';
 
 export default async function removeMode(packageName) {
   console.log(chalk.green.bold(`Removing ohif-mode ${packageName}...`));
-
-  async function removeModeFromConfigFile() {
-    const pluginConfig = readPluginConfigFile();
-
-    // Note: if file is not found, nothing to remove.
-    if (pluginConfig) {
-      removeModeFromConfig(pluginConfig, { packageName });
-      writePluginConfigFile(pluginConfig);
-    }
-  }
-
-  async function findOhifExtensionsToRemove(yarnInfo) {
-    const pluginConfig = readPluginConfigFile();
-
-    if (!pluginConfig) {
-      // No other modes or extensions, no action item.
-      return [];
-    }
-
-    const { modes, extensions } = pluginConfig;
-
-    const registeredExtensions = extensions.map(
-      extension => extension.packageName
-    );
-    // TODO this is not a function
-    const ohifExtensionsOfMode = Object.keys(
-      yarnInfo.peerDependencies
-    ).filter(peerDependency => registeredExtensions.includes(peerDependency));
-
-    const ohifExtensionsUsedInOtherModes = ohifExtensionsOfMode.map(
-      packageName => {
-        return {
-          packageName,
-          used: false,
-        };
-      }
-    );
-
-    // Check if other modes use each extension used by this mode
-    const otherModes = modes.filter(mode => mode.packageName !== yarnInfo.name);
-
-    for (let i = 0; i < otherModes.length; i++) {
-      const mode = otherModes[i];
-      const yarnInfo = await getYarnInfo(mode.packageName);
-
-      const peerDependencies = yarnInfo.peerDependencies;
-
-      if (!peerDependencies) {
-        continue;
-      }
-
-      for (let j = 0; j < ohifExtensionsUsedInOtherModes.length; j++) {
-        const ohifExtension = ohifExtensionsUsedInOtherModes[j];
-        if (ohifExtension.used) {
-          // Already accounted that we can't delete this, so don't waste effort
-          return;
-        }
-
-        if (Object.keys(peerDependencies).includes(ohifExtension.packageName)) {
-          ohifExtension.used = true;
-        }
-      }
-    }
-
-    // Return list of now unused extensions
-    const ohifExtensionsToRemove = ohifExtensionsUsedInOtherModes
-      .filter(ohifExtension => !ohifExtension.used)
-      .map(ohifExtension => ohifExtension.packageName);
-
-    return ohifExtensionsToRemove;
-  }
 
   const tasks = new Listr(
     [
@@ -105,7 +33,7 @@ export default async function removeMode(packageName) {
       {
         title: 'Detecting extensions that can be removed...',
         task: async ctx => {
-          ctx.ohifExtensionsToRemove = await findOhifExtensionsToRemove(
+          ctx.ohifExtensionsToRemove = await findOhifExtensionsToRemoveAfterRemovingMode(
             ctx.yarnInfo
           );
         },
